Input with storybook utils
