+++ web/src/stories/Form/Input.stories.ts
@@ -1,19 +1,5 @@
 import type { StoryObj, Meta } from '@storybook/html'
-
-type Props = HTMLInputElement & {
-  fullWidth: boolean
-  placeholder: string
-}
-
-export const createInput = ({ fullWidth, ...props }: Props) => {
-  const input = document.createElement('input')
-
-  for (const [key, value] of Object.entries(props)) {
-    input.setAttribute(key, value?.toString() || '')
-  }
-  fullWidth && input.classList.add('full-width')
-  return input
-}
+import { createInput, Input as Props } from '../utils'
 
 /**
  * Standard input field
