+++ web/src/stories/types.ts
@@ -1,1 +1,6 @@
 export type Link = { href: string; text: string }
+
+export type Input = HTMLInputElement & {
+  fullWidth: boolean
+  placeholder: string
+}
