+++ web/src/stories/utils.ts
@@ -1,0 +1,13 @@
+import { Input } from './types'
+
+export const createInput = ({ fullWidth, ...props }: Input) => {
+  const input = document.createElement('input')
+
+  for (const [key, value] of Object.entries(props)) {
+    input.setAttribute(key, value?.toString() || '')
+  }
+  fullWidth && input.classList.add('full-width')
+  return input
+}
+
+export * from './types'
